backend/requirements.txt
fastapi[standard]>=0.110.0
uvicorn>=0.29.0
supabase>=2.4.0
google-cloud-aiplatform>=1.43.0
pydantic>=2.6.3
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
pyjwt>=2.8.0
